Get JRuby extension to work.
